BUG: JS error if user sets theme to empty string or something other than one of the library choices
selectBoxIt.js defaults to "bootstrap" if you don't set a
theme, and if you leave the theme option empty or try to
set it to "none" it crashes when trying to go through
_addClasses.  Looks like there is no default object that
_addClasses should consume if a user doesn't use one of
the framework theme option.

I'll leave the question about whether it should be
defaulting to no theme rather than "bootstrap" to you,
but there probably needs to be a fallback "obj" object
created if _addClass() is called without an obj as in the
else case when setting the theme.