Added support for 'copyClasses' option
Added support for the new 'copyClasses' option.  This option allows a
user to specify if they would like the native select box class(s) to be
copied over to the drop down container (which holds both the drop down
button and drop down list) (with value 'container') or the drop down
button (with the default value 'button').  The option also takes the
value 'none' to indicate that native classes should not be copied to
the SelectBoxIt object.